Added initial Blood and Fecal Testing, removal
Two state charts were added to track the test status of individual cows
and to allow for a delay of 14 months for Fecal testing results and 2
weeks for Blood test results. These state charts are
fecalTestingStateChart and bloodTestingStateChart.

Two action charts were also added. It is simple to follow the action
charts by their logic, block names and descriptions. Please open an issue
and put Alex as the owner of the issue if there is a problem with the
logic.

Variables were added to main to track cows removed by true positive and
false positive testing results from each testing type. I also added an
OptionList of TestingType because I intend to use it later for function
logic.

NOTE: No significant change to model runtime lag after adding the action
charts and state charts

TODO Add a timeplot to track cows removed via positive test results
TODO Add a feature to stop testing after we are quite sure the herd is
disease free