Fixed object names, cleaned view, added scenarios
I cleaned up the model in main and Cow. I got rid of unnecessary parameter
groups, named all parameter groups, and then moved the parameter groups to
the top left quadrant. I named all charts and put them in the top right
quadrant. In Cow, I moved all parameter groups to the left side and I also
fixed the names of some conditions and code blocks in the actions charts.

I added multiple scenarios. The scenario with 5000 initial cows was
changed to veryHighPopulation, the scenario with 1000 population was named
highPopulation, and I created a scenario called normalPopulation with a
population of 100. In addition, I added a scenario called
sensitivityAnalysis which can be used for the sensitivity analysis section
of the report. I would suggest you save the scenario under a specific name
if it ends up being significant enough to write about, and then Cheryl and
Nate can look into it.