Fixed Parameter issues and simulation param values
BloodTestSubclinicalSensitivity and BloodTestClinicalSensitivity had swapped values. All simulations had altered values for InitialInfectedPercentage, for example, and so I set them all to be the default unless a different value was required for the scenario. FrequencyOfInfectiveContact was set to 7.5 for all simulations instead of 7.1.

Please make asure the model is stil properly calibrated. These issues have been present for weeks.